fix: correction storybook
