+++ src/pages/people/outlets/color/palette/form/RenderContentFormPalette/stories/RenderContentFormPalette.stories.tsx
@@ -1,21 +1,18 @@
 import { StoryFn } from "@storybook/react";
 import { BrowserRouter } from "react-router-dom";
-import { presente, inube, Stack } from "@inube/design-system";
-
-import {
-  RenderStrokesWithSpinnerForm,
-  RenderStrokesWithSpinnerFormProps,
-} from ".";
-import { strokesFormsConfig } from "../../config/Strokes.config";
+import { presente, Stack, inube } from "@inube/design-system";
+import { RenderContentFormPalette, RenderContentFormPaletteProps } from "..";
 import { TokenContext } from "@src/context/TokenContext";
+import { paletteTabsConfig } from "../../../config/paletteTabs.config";
+import { props } from "./props";
 
 const story = {
-  components: [RenderStrokesWithSpinnerForm],
-  title:
-    "layouts/people/outlets/color/strokes/form/RenderStrokesWithSpinnerForm",
+  components: [RenderContentFormPalette],
+  title: "layouts/people/outlets/color/palette/form/RenderContentFormPalette",
   parameters: {
     layout: "fullscreen",
   },
+  argTypes: props,
   decorators: [
     (Story: StoryFn) => (
       <BrowserRouter>
@@ -29,16 +26,17 @@ const story = {
   ],
 };
 
-const Default = (args: RenderStrokesWithSpinnerFormProps) => {
+const Default = (args: RenderContentFormPaletteProps) => {
   return (
     <Stack padding="s300" direction="column" gap={inube.spacing.s400}>
-      <RenderStrokesWithSpinnerForm {...args} formType="spinner" />
+      <RenderContentFormPalette {...args} />
     </Stack>
   );
 };
 
 Default.args = {
-  strokesConfig: { spinner: strokesFormsConfig.spinner },
+  paletteTabsConfig: { neutral: paletteTabsConfig.neutral },
+  formType: "neutral",
 };
 
 export default story;
