+++ src/pages/people/outlets/color/palette/form/RenderContentFormPalette/stories/props.ts
@@ -1,0 +1,10 @@
+const props = {
+  formType: {
+    options: ["neutral", "neutralAlpha"],
+    control: { type: "select" },
+    table: {
+      defaultValue: { summary: "neutral" },
+    },
+  },
+};
+export { props };
