+++ src/pages/people/outlets/color/strokes/form/RenderStrokesContentForm/stories/RenderStrokesContentForm.stories.tsx
@@ -1,21 +1,18 @@
 import { StoryFn } from "@storybook/react";
 import { BrowserRouter } from "react-router-dom";
-import { presente, inube, Stack } from "@inube/design-system";
-
-import {
-  RenderStrokesWithSpinnerForm,
-  RenderStrokesWithSpinnerFormProps,
-} from ".";
-import { strokesFormsConfig } from "../../config/Strokes.config";
+import { presente, Stack, inube } from "@inube/design-system";
+import { RenderStrokesContentForm, RenderStrokesContentFormProps } from "..";
 import { TokenContext } from "@src/context/TokenContext";
+import { strokesFormsConfig } from "../../../config/Strokes.config";
+import { props } from "./props";
 
 const story = {
-  components: [RenderStrokesWithSpinnerForm],
-  title:
-    "layouts/people/outlets/color/strokes/form/RenderStrokesWithSpinnerForm",
+  components: [RenderStrokesContentForm],
+  title: "layouts/people/outlets/color/strokes/form/RenderStrokesContentForm",
   parameters: {
     layout: "fullscreen",
   },
+  argTypes: props,
   decorators: [
     (Story: StoryFn) => (
       <BrowserRouter>
@@ -29,16 +26,17 @@ const story = {
   ],
 };
 
-const Default = (args: RenderStrokesWithSpinnerFormProps) => {
+const Default = (args: RenderStrokesContentFormProps) => {
   return (
     <Stack padding="s300" direction="column" gap={inube.spacing.s400}>
-      <RenderStrokesWithSpinnerForm {...args} formType="spinner" />
+      <RenderStrokesContentForm {...args} />
     </Stack>
   );
 };
 
 Default.args = {
-  strokesConfig: { spinner: strokesFormsConfig.spinner },
+  strokesConfig: { ...strokesFormsConfig },
+  formType: "primary",
 };
 
 export default story;
