+++ src/pages/people/outlets/color/strokes/form/RenderStrokesContentForm/stories/props.ts
@@ -1,0 +1,20 @@
+const props = {
+  formType: {
+    options: [
+      "primary",
+      "error",
+      "warning",
+      "success",
+      "information",
+      "help",
+      "dark",
+      "gray",
+      "light",
+    ],
+    control: { type: "select" },
+    table: {
+      defaultValue: { summary: "primary" },
+    },
+  },
+};
+export { props };
