+++ src/pages/people/outlets/color/strokes/form/RenderStrokesWithLinkForm/RenderStrokesWithLinkForm.stories.tsx
@@ -1,9 +1,9 @@
 import { StoryFn } from "@storybook/react";
 import { BrowserRouter } from "react-router-dom";
-import { presente, inube, Stack } from "@inube/design-system";
-
-import { RenderStrokesWithLinkForm, RenderStrokesWithLinkFormProps } from ".";
+import { presente, Stack, inube } from "@inube/design-system";
+import { TokenContext } from "@src/context/TokenContext";
 import { strokesFormsConfig } from "../../config/Strokes.config";
+import { RenderStrokesWithLinkForm, RenderStrokesWithLinkFormProps } from ".";
 
 const story = {
   components: [RenderStrokesWithLinkForm],
@@ -14,52 +14,27 @@ const story = {
   decorators: [
     (Story: StoryFn) => (
       <BrowserRouter>
-        <Story />
+        <TokenContext.Provider
+          value={{ token: presente, handleSubmit: () => {} }}
+        >
+          <Story />
+        </TokenContext.Provider>
       </BrowserRouter>
     ),
   ],
 };
 
-const themeMap = {
-  presente: presente,
-  inube: inube,
-};
 const Default = (args: RenderStrokesWithLinkFormProps) => {
-  const selectedTheme = themeMap[args.token as keyof typeof themeMap];
-
   return (
-    <Stack padding="s300" direction="column" gap={selectedTheme.spacing.s400}>
-      <RenderStrokesWithLinkForm {...args} token={selectedTheme} />
+    <Stack padding="s300" direction="column" gap={inube.spacing.s400}>
+      <RenderStrokesWithLinkForm {...args} formType="link" />
     </Stack>
   );
 };
 
 Default.args = {
-  formType: "primary",
-  handleSubmit: () => {},
-  token: "presente",
-  strokesConfig: strokesFormsConfig,
+  strokesConfig: { link: strokesFormsConfig.link },
 };
-
-Default.argTypes = {
-  token: {
-    options: ["presente", "inube"],
-    control: { type: "select" },
-    description: "the theme that it be use to render",
-    table: {
-      defaultValue: { summary: "inube" },
-    },
-  },
-  formType: {
-    options: Object.keys(strokesFormsConfig),
-    control: { type: "select" },
-    description: "the form that it'll be render",
-    table: {
-      defaultValue: { summary: "primary" },
-    },
-  },
-};
-
 export default story;
 
 export { Default };
